coemnzando con la creacion de series
