--- src/main/java/com/nanoka/restaurant_api/receipt/adapters/input/rest/model/response/OrderDetailResponse.java
@@ -1,17 +1,0 @@
-package com.nanoka.restaurant_api.receipt.adapters.input.rest.model.response;
-
-import lombok.*;
-
-import java.math.BigDecimal;
-
-@Builder
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class OrderDetailResponse {
-    private Long id;
-    private ProductResponse product;
-    private BigDecimal unitPrice;
-    private int quantity;
-}
