--- src/main/java/com/nanoka/restaurant_api/receipt/adapters/input/rest/model/response/ProductResponse.java
@@ -1,13 +1,0 @@
-package com.nanoka.restaurant_api.receipt.adapters.input.rest.model.response;
-
-import lombok.*;
-
-@Builder
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class ProductResponse {
-    private Long id;
-    private String name;
-}
