+++ src/main/java/com/nanoka/restaurant_api/receipt/application/ports/output/ReceiptPersistencePort.java
@@ -1,6 +1,6 @@
 package com.nanoka.restaurant_api.receipt.application.ports.output;
 
-import com.nanoka.restaurant_api.receipt.model.Receipt;
+import com.nanoka.restaurant_api.receipt.domain.model.Receipt;
 
 import java.util.List;
 import java.util.Optional;
