+++ src/main/java/com/nanoka/restaurant_api/receipt/application/services/ReceiptService.java
@@ -2,7 +2,7 @@ package com.nanoka.restaurant_api.receipt.application.services;
 
 import com.nanoka.restaurant_api.receipt.application.ports.input.ReceiptServicePort;
 import com.nanoka.restaurant_api.receipt.application.ports.output.ReceiptPersistencePort;
-import com.nanoka.restaurant_api.receipt.model.Receipt;
+import com.nanoka.restaurant_api.receipt.domain.model.Receipt;
 import com.nanoka.restaurant_api.util.ErrorCatelog;
 import com.nanoka.restaurant_api.util.exceptions.NotFoundException;
 import lombok.RequiredArgsConstructor;
