+++ src/main/java/com/nanoka/restaurant_api/receipt/infrastructure/adapters/input/rest/model/request/ReceiptDetailCreateRequest.java
@@ -1,0 +1,13 @@
+package com.nanoka.restaurant_api.receipt.infrastructure.adapters.input.rest.model.request;
+
+import lombok.*;
+
+@Builder
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class ReceiptDetailCreateRequest {
+    private Long orderDetailId;
+    private int quantity;
+}
