+++ src/main/java/com/nanoka/restaurant_api/serie/application/ports/output/SeriePersistencePort.java
@@ -1,0 +1,5 @@
+package com.nanoka.restaurant_api.serie.application.ports.output;
+
+public class SeriePersistencePort {
+
+}
